KSCU-74: third-wave clean-up (ex: remove unneeded variable declarations, remove old comments, move recreateCartHelpers into appropriate folder, adjust syntax, etc.)
