+++ cartridges/int_klaviyo_core/cartridge/scripts/klaviyo/checkoutHelpers.js
@@ -6,7 +6,7 @@ var klaviyoUtils = require('*/cartridge/scripts/klaviyo/utils');
 var startedCheckoutData = require('*/cartridge/scripts/klaviyo/eventData/startedCheckout');
 
 /* API Includes */
-var basketMgr = require('dw/order/BasketMgr');
+var BasketMgr = require('dw/order/BasketMgr');
 
 
 // beginCheckout set to true if this is being called from the first view / step of checkout.
@@ -33,12 +33,10 @@ function startedCheckoutHelper(beginCheckout, customerEmail) {
             var exchangeID = klaviyoUtils.getKlaviyoExchangeID();
             var dataObj, serviceCallResult, currentBasket;
             var isKlDebugOn = request.httpParameterMap.kldebug.booleanValue;
-
-            currentBasket = basketMgr.getCurrentBasket();
+            currentBasket = BasketMgr.getCurrentBasket();
 
             if ( customerEmail ) {
-
-                if (currentBasket && currentBasket.getProductLineItems().toArray().length) { //TODO: is there a property for isEmpty on basket object?
+                if (currentBasket && currentBasket.getProductLineItems().toArray().length) {
                     dataObj = startedCheckoutData.getData(currentBasket);
                     serviceCallResult = klaviyoUtils.trackEvent(exchangeID, dataObj, klaviyoUtils.EVENT_NAMES.startedCheckout, customerEmail);
                     if (isKlDebugOn) {
@@ -50,7 +48,6 @@ function startedCheckoutHelper(beginCheckout, customerEmail) {
                 session.privacy.klaviyoCheckoutTracked = true;
 
             } else {
-                // TODO: return this value so it can be added to pdict in sitegen template call, and add return false at the bottom of this function
                 returnObj.klid = klaviyoUtils.getProfileInfo();
             }
 
@@ -63,13 +60,11 @@ function startedCheckoutHelper(beginCheckout, customerEmail) {
 
 
 function getEmailFromBasket() {
-    var currentBasket = basketMgr.getCurrentBasket();
+    var currentBasket = BasketMgr.getCurrentBasket();
     if(currentBasket && currentBasket.customerEmail) {
         return currentBasket.customerEmail;
     }
-    // if(currentBasket && currentBasket.customer && currentBasket.customer.profile && currentBasket.customer.profile.email) {
-    //     return currentBasket.customer.profile.email;
-    // }
+
     return false;
 }
